Move model-utils tests and add io.github.ygojson.model.utils.serialization package
