+++ model/src/test/java/io/github/ygojson/model/utils/acceptancetest/SerializationAcceptanceTest.java
@@ -1,4 +1,4 @@
-package io.github.ygojson.model.utils.schema.acceptance;
+package io.github.ygojson.model.utils.acceptancetest;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import io.github.ygojson.model.utils.JsonUtils;
+import io.github.ygojson.model.utils.serialization.JsonUtils;
 
 public class SerializationAcceptanceTest {
