+++ model/src/test/java/io/github/ygojson/model/utils/data/definition/localization/CardLocalizedDataUnitTest.java
