+++ model/src/test/java/io/github/ygojson/model/utils/data/utils/serialization/JsonUtilsUnitTest.java
@@ -1,4 +1,4 @@
-package io.github.ygojson.model.utils;
+package io.github.ygojson.model.utils.data.utils.serialization;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -9,6 +9,8 @@ import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+
+import io.github.ygojson.model.utils.serialization.JsonUtils;
 
 class JsonUtilsUnitTest {
