+++ model/src/test/resources/io/github/ygojson/model/utils/acceptancetest/.gitignore
