+++ model/src/test/resources/io/github/ygojson/model/utils/acceptancetest/SerializationAcceptanceTest.testDataModelSerialization.Card.approved.json
