+++ model/src/test/resources/io/github/ygojson/model/utils/acceptancetest/SerializationAcceptanceTest.testDataModelSerialization.Set.approved.json
