+++ tools/src/test/java/io/github/ygojson/tools/dataprovider/impl/yugipedia/acceptance/YugipediaSetMapperTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.mapstruct.factory.Mappers;
 
 import io.github.ygojson.model.data.Set;
-import io.github.ygojson.model.utils.JsonUtils;
+import io.github.ygojson.model.utils.serialization.JsonUtils;
 import io.github.ygojson.tools.dataprovider.impl.yugipedia.YugipediaTestData;
 import io.github.ygojson.tools.dataprovider.impl.yugipedia.mapper.YugipediaSetMapper;
 import io.github.ygojson.tools.dataprovider.impl.yugipedia.mapper.wikitext.InfoboxSetMapper;
